requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
distro==1.8.0
dnspython==2.4.2
email-validator==2.0.0.post2
Faker==19.10.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==1.11.1
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
flask-marshmallow==0.10.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-SSLify==0.1.5
Flask-WTF==1.2.1
frozenlist==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
multidict==6.0.4
numpy==1.26.1
openai==1.2.3
packaging==23.2
Pillow==10.1.0
pipdeptree==2.13.0
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sseclient-py==1.8.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
validate-email==1.3
Werkzeug==3.0.1
WTForms==3.1.0
yarl==1.9.2
